requirements.txt
coverage==4.2
dj-database-url==0.4.1
Django==1.10.1
django-allauth==0.27.0
django-cleanup==0.4.2
django-nose==1.4.4
django-rest-auth==0.8.1
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
funcsigs==1.0.2
future==0.15.2
gunicorn==19.6.0
mock==2.0.0
nose==1.3.7
oauthlib==2.0.0
pbr==1.10.0
Pillow==3.3.1
psycopg2==2.6.2
PyJWT==1.4.2
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
whitenoise==3.2.1
